refactor: organize net module implementation
